api/requirements.txt
flask
pandas
mysql-connector-python
PyMySQL
fire
